feat(profile): phase 2 — settings hub revamp, photo sheet, notifications rehouse, typed delete
Rewrites the settings hub with grouped cards, moves the photo editor
inline via a MobileBottomSheet, rehouses notifications under the
settings shell, and replaces the double-click delete with a typed
confirmation dialog.

**Hub rewrite** (`profile.index.tsx`)
- Grouped cards using `Card` + ui-kit primitives: Profile / Account /
  Preferences / Your data / Danger zone.
- Profile card: avatar (click → photo sheet) + inline name/username
  form (Conform + Zod, async username uniqueness preserved).
- Account card: Email / Password / 2FA rows with explicit aria-labels.
- Connections stub removed from the hub — no OAuth yet, dead link.
- Danger zone is visually separate with destructive styling.

**Photo bottom sheet** (`__profile-photo-sheet.tsx`)
- Extracts the cropper UI from the old `/settings/profile/photo` page
  into a `MobileBottomSheet` triggered from the avatar. Uses
  `profile-photo-utils.ts` (createCroppedFile / useObjectUrl) unchanged.
- `react-easy-crop` is lazy-loaded via `React.lazy()` + Suspense —
  its module-level `window` access crashes SSR when pulled into the
  settings hub bundle.
- Submits by building FormData manually in `handleSave` and firing a
  fetcher to the existing `/settings/profile/photo` action-only route.
  Wrapping the file input in a form + Conform caused the dialog to
  unmount on change (Radix focus + Conform revalidation race).

**Photo route** (`profile.photo.tsx`)
- Stripped to an action-only resource route: loader redirects GET to
  `/settings/profile`, action handles upload + delete with the same
  Zod schema as before. No more default component.

**Danger zone typed confirm** (`__danger-zone-delete-dialog.tsx`)
- Replaces the one-click double-check with a dialog that requires the
  user to type their exact username before enabling the delete button.
  Wired to the existing `deleteDataAction` intent on the hub.

**Notifications rehouse**
- Moves `notifications.tsx` → `profile.notifications.tsx` so it inherits
  the settings shell from `profile.tsx`. All internal references and the
  `getPreferenceManagementUrl` helper now emit the new path.
- Leaves `notifications.tsx` in place as a 301-style redirect for
  existing magic links that still point to the old path.

**Layout cleanup** (`profile.tsx`)
- Drops the breadcrumb chain rooted at `/users/wade`. Settings is its
  own top-level destination; each sub-route owns its own heading.

**Tests**
- `settings-profile.test.ts`: photo test now opens the sheet from the
  avatar and verifies the dialog + file input + disabled Save. New
  typed-confirm delete test exercises the wrong-value / right-value
  gate on the new dialog.
- `settings-notifications.test.ts`: all request URLs updated to the
  new path.
- `profile.notifications.test.tsx`: request URLs + import path updated.

Typecheck clean, 0 lint errors, 505/505 unit, 13/13 relevant e2e.
