+++ app/routes/settings+/__danger-zone-delete-dialog.tsx
@@ -1,0 +1,102 @@
+import { useEffect, useState } from 'react';
+import { useFetcher } from 'react-router';
+import { Button } from '#app/components/ui/button.tsx';
+import {
+  Dialog,
+  DialogContent,
+  DialogDescription,
+  DialogFooter,
+  DialogHeader,
+  DialogTitle,
+} from '#app/components/ui/dialog.tsx';
+import { Icon } from '#app/components/ui/icon.tsx';
+import { Input } from '#app/components/ui/input.tsx';
+import { Label } from '#app/components/ui/label.tsx';
+import { StatusButton } from '#app/components/ui/status-button.tsx';
+
+type DangerZoneDeleteDialogProps = Readonly<{
+  username: string;
+  intent: string;
+}>;
+
+// Typed-confirmation dialog for the "delete all your data" action. Replaces
+// the old double-click button — a single tap is too easy for an irreversible
+// `prisma.user.delete`. The delete button only enables when the user types
+// their exact username. Submits to the hub's action with the provided intent
+// string.
+export function DangerZoneDeleteDialog({
+  username,
+  intent,
+}: DangerZoneDeleteDialogProps) {
+  const fetcher = useFetcher();
+  const [open, setOpen] = useState(false);
+  const [typed, setTyped] = useState('');
+  const isPending = fetcher.state !== 'idle';
+  const matches = typed.trim() === username;
+
+  useEffect(() => {
+    if (!open) setTyped('');
+  }, [open]);
+
+  return (
+    <>
+      <Button
+        type="button"
+        variant="destructive"
+        onClick={() => setOpen(true)}
+      >
+        <Icon name="trash">Delete all your data</Icon>
+      </Button>
+
+      <Dialog open={open} onOpenChange={setOpen}>
+        <DialogContent className="sm:max-w-md">
+          <DialogHeader>
+            <DialogTitle>Delete your account?</DialogTitle>
+            <DialogDescription>
+              This permanently removes your profile, wishlist, friends, and
+              every group you own. It cannot be undone.
+            </DialogDescription>
+          </DialogHeader>
+
+          <fetcher.Form method="POST" className="flex flex-col gap-3">
+            <Label htmlFor="danger-zone-confirm" className="text-sm">
+              Type{' '}
+              <span className="font-mono font-semibold text-foreground">
+                {username}
+              </span>{' '}
+              to confirm.
+            </Label>
+            <Input
+              id="danger-zone-confirm"
+              autoComplete="off"
+              spellCheck={false}
+              value={typed}
+              onChange={(event) => setTyped(event.target.value)}
+              placeholder={username}
+            />
+
+            <DialogFooter className="mt-2 flex flex-col-reverse gap-2 sm:flex-row sm:justify-end">
+              <Button
+                type="button"
+                variant="outline"
+                onClick={() => setOpen(false)}
+              >
+                Cancel
+              </Button>
+              <StatusButton
+                type="submit"
+                name="intent"
+                value={intent}
+                variant="destructive"
+                disabled={!matches || isPending}
+                status={isPending ? 'pending' : 'idle'}
+              >
+                <Icon name="trash">Delete my account</Icon>
+              </StatusButton>
+            </DialogFooter>
+          </fetcher.Form>
+        </DialogContent>
+      </Dialog>
+    </>
+  );
+}
